more component consolidation + installed eslint
