+++ client/src/components/VerticalNav.jsx
@@ -1,19 +1,19 @@
 import React from 'react';
-import VerticalNavItem from './VerticalNavItem.jsx';
 import { faPlus, faFaceSmile, faBook } from '@fortawesome/free-solid-svg-icons';
 import styles from '../stylesheets/vertical-nav.css';
 import { useNavigate } from 'react-router-dom';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
 const VerticalNav = props => {
 
   const navigate = useNavigate();
   const onNavItemClick = (event, btnName) => {
     switch (btnName) {
-    case 'add-restaurant':
-      navigate('/new-restaurant');
-      break;
-    default:
-      navigate('/login');
+      case 'add-restaurant':
+        navigate('/new-restaurant');
+        break;
+      default:
+        navigate('/login');
     }
   };
 
@@ -34,5 +34,16 @@ const VerticalNav = props => {
     </div>
   );
 };
+
+const VerticalNavItem = props => {
+  return (
+    <button
+      className="vertical-nav-btn"
+      onClick={(event) => props.onClickHandler(event, props.btnName)}>
+      <FontAwesomeIcon icon={props.iconName} />
+    </button>
+  );
+};
+
 
 export default VerticalNav;
